refactor: deplacement des composants dans un dossier routes
